+++ homematicip/aio/class_maps.py
@@ -16,6 +16,7 @@ TYPE_CLASS_MAP = {
     DeviceType.BRAND_SWITCH_MEASURING: AsyncBrandSwitchMeasuring,
     DeviceType.BRAND_SWITCH_NOTIFICATION_LIGHT: AsyncBrandSwitchNotificationLight,
     DeviceType.BRAND_WALL_MOUNTED_THERMOSTAT: AsyncWallMountedThermostatPro,
+    DeviceType.DIN_RAIL_SWITCH_4: AsyncDinRailSwitch4,
     DeviceType.FLOOR_TERMINAL_BLOCK_10: AsyncFloorTerminalBlock10,
     DeviceType.FLOOR_TERMINAL_BLOCK_12: AsyncFloorTerminalBlock12,
     DeviceType.FLOOR_TERMINAL_BLOCK_6: AsyncFloorTerminalBlock6,